Refactoring helper and jobs
