+++ spree_retailops.gemspec
@@ -18,7 +18,7 @@ Gem::Specification.new do |s|
   s.require_path = 'lib'
   s.requirements << 'none'
 
-  #s.add_dependency 'spree_core', '~> 2.2.1'
+  s.add_dependency 'activejob_backport'  # if RAILS_VERSION < "4.2"
 
   s.add_development_dependency 'capybara', '~> 2.1'
   s.add_development_dependency 'coffee-rails'
@@ -30,7 +30,5 @@ Gem::Specification.new do |s|
   s.add_development_dependency 'selenium-webdriver'
   s.add_development_dependency 'simplecov'
   s.add_development_dependency 'sqlite3'
-  s.add_development_dependency 'spree'
-  s.add_development_dependency 'activejob_backport'  # if RAILS_VERSION < "4.2"
-  s.add_development_dependency 'pry-byebug'
+  s.add_development_dependency 'spree_core', '~> 2.2.1'
 end
